small corrections

+++ algorithms/maze_generator_recursive_backtracking.c
@@ -27,8 +27,8 @@ inline const u64 get_next(node* maze, u64 node_idx, u32 height, u32 width) {
     int i = rand() & 3;
 
     for (uint attempt = 0; attempt < 4; ++attempt) {
-        uint x = maze[node_idx].x + pos_change[i][0];
-        uint y = maze[node_idx].y + pos_change[i][1];
+        u32 x = maze[node_idx].x + pos_change[i][0];
+        u32 y = maze[node_idx].y + pos_change[i][1];
 
         i = (i + 1) & 3;
 
@@ -46,8 +46,8 @@ inline const u32 get_next(node* maze, u32 node_idx, u32 height, u32 width) {
     int i = rand() & 3;
 
     for (uint attempt = 0; attempt < 4; ++attempt) {
-        uint x = maze[node_idx].x + pos_change[i][0];
-        uint y = maze[node_idx].y + pos_change[i][1];
+        u32 x = maze[node_idx].x + pos_change[i][0];
+        u32 y = maze[node_idx].y + pos_change[i][1];
 
         i = (i + 1) & 3;
 
@@ -77,20 +77,20 @@ char* generate_maze(u32 height, u32 width) {
     // height
 
     struct node* pre_maze = (struct node*) malloc(height * width / 4 * sizeof(struct node));
-    for (uint i = 0; i < height * width / 4; ++i) {
+    for (u64 i = 0; i < height * width / 4; ++i) {
         pre_maze[i].pred_node = -2;
         pre_maze[i].x = i / (width / 2);
         pre_maze[i].y = i % (width / 2);
     }
 
-    uint length = (width + 2) * (height + 1);
+    u64 length = (width + 2) * (height + 1);
 
     char* maze = (char*) malloc((length));
-    for (uint i = 0; i < length; ++i) {
+    for (u64 i = 0; i < length; ++i) {
         maze[i] = '#';
     }
 
-    for (uint i = width + 1; i < length; i += width + 2) {
+    for (u64 i = width + 1; i < length; i += width + 2) {
         maze[i] = '\n';
     }
 
@@ -115,8 +115,8 @@ char* generate_maze(u32 height, u32 width) {
          next_node= get_next(pre_maze, current_node, height / 2, width / 2);
 
         if (next_node == null_node) {
-            uint x = pre_maze[current_node].x;
-            uint y = pre_maze[current_node].y;
+            u32 x = pre_maze[current_node].x;
+            u32 y = pre_maze[current_node].y;
 
             uint off_x = (x <= pre_maze[pre_maze[current_node].pred_node].x) + (x < pre_maze[pre_maze[current_node].pred_node].x);
             uint off_y = (y <= pre_maze[pre_maze[current_node].pred_node].y) + (y < pre_maze[pre_maze[current_node].pred_node].y);
